refactor(team): split headless_codex.go into themed sibling files (C10)
Mechanical file split. The 1914-line headless_codex.go owned every
layer of headless dispatch (entry points, types, queue management,
codex CLI invocation, recovery, completion checks, env builders).
Splitting into themed siblings preserves all behaviour but gives
each layer its own file boundary so navigation and review match the
layered responsibilities.

New files:

* headless_codex_queue.go (581 lines, new) — per-slug queue
  management. enqueueHeadlessCodexTurn[Record], dedup/replace,
  lead-wake heuristics, worker lifecycle (spawn/stop/runQueue),
  per-turn lifecycle (begin/finish, wakeLeadAfterSpecialist,
  latestLeadWakeTaskAction), and the timeout/stale-cancel helpers.

* headless_codex_runner.go (705 lines, new) — the codex-CLI
  invocation half. Hosts runHeadlessCodexTurn (the 220-line method
  that builds the codex command line, pipes prompt over stdin,
  parses the JSON event stream, and surfaces the result), plus its
  env/auth/workspace builders (buildHeadlessCodexEnv,
  buildCodexOfficeConfigOverrides, preflightHeadlessCodexAuth,
  dumpHeadlessCodexInvocation, headlessCodex{Home,Global,Runtime}Dir,
  prepareHeadlessCodexHome, copyHeadlessCodexHomeFile,
  normalizeHeadlessWorkspaceDir, headlessCodexWorkspaceCacheDir,
  headlessTaskWorkspaceDir, headlessCodexNeedsDangerousBypass,
  buildHeadlessCodexPrompt, isCodexAuthError) and the toml/env
  utility helpers (tomlQuote, tomlStringArray, setEnvValue,
  stripEnvKeys).

headless_codex.go (664 lines, was 1914) — keeps the dispatch entry
points (defaultHeadlessCodexRunTurn, headlessCodexRunTurn,
launchHeadlessCodex), package vars + types (headlessCodexTurn,
headlessCodexActiveTurn, headlessCodexWorkspaceStatusSnapshot
override seam), and the recovery / durability-check / completion
methods (taskHasDurableCompletionState,
headlessTurnCompletedDurably, taskHasExternalWorkflowEvidenceSince,
recoverTimedOutHeadlessTurn, recoverFailedHeadlessTurn,
timedOutTurnAlreadyRecovered, isDurabilityFailure, retry-prompt
helpers, and the agent-posting heuristics).

Imports trimmed per file to only what each one uses (bufio + io
moved entirely to runner; errors + sync/atomic dropped from queue;
agent + gitexec dropped from headless_codex.go).

No new tests — pure file split. Existing test suite (32 packages
green) is the safety net. Each function appears in exactly one
file before, exactly one file after, with the same body.

Stacked on refactor/launcher-c9-wrapper-sweep (C9).
